// src/app/models/supplier.ts
export class Supplier {
    id: number;
    tenent_code: string;
    name: string;
    display_name: string;
    primary_user_id: number;
    type: number;
    baseurl: string;
    primary_user_name: string;
    primary_user_mobie: string;
    primary_user_email: string;
    services: string;
}